+++ docs/api/version-history.md
@@ -32,6 +32,8 @@ keywords: "API, Docker, rcli, REST, documentation"
   the multi-platform image.
 * `POST /containers/create` now takes `Options` as part of `HostConfig.Mounts.TmpfsOptions` to set options for tmpfs mounts.
 * `POST /services/create` now takes `Options` as part of `ContainerSpec.Mounts.TmpfsOptions`, to set options for tmpfs mounts.
+* `GET /events` now supports image `create` event that is emitted when a new
+  image is built regardless if it was tagged or not.
 
 ### Deprecated Config fields in `GET /images/{name}/json` response